blok edit sampai stok

+++ resources/views/gudang/_modal.blade.php
@@ -63,6 +63,7 @@
   let modalBody = $('#crudModal').find('.modal-body').html()
 
   let dataMaxLength = []
+  var data_id
 
   $(document).ready(function() {
     $('#btnSubmit').click(function(event) {
@@ -176,6 +177,7 @@
     setFormBindKeys(form)
 
     activeGrid = null
+    data_id = $('#crudForm').find('[name=id]').val();
 
     form.find('#btnSubmit').prop('disabled', false)
     if (form.data('action') == "view") {
@@ -188,8 +190,38 @@
 
   $('#crudModal').on('hidden.bs.modal', () => {
     activeGrid = '#jqGrid'
+    removeEditingBy(data_id)
     $('#crudModal').find('.modal-body').html(modalBody)
   })
+  function removeEditingBy(id) {
+    $.ajax({
+      url: `{{ config('app.api_url') }}bataledit`,
+      method: 'POST',
+      dataType: 'JSON',
+      headers: {
+        Authorization: `Bearer ${accessToken}`
+      },
+      data: {
+        id: id,
+        aksi: 'BATAL',
+        table: 'gudang'
+        
+      },
+      success: response => {
+        $("#crudModal").modal("hide")
+      },
+      error: error => {
+        if (error.status === 422) {
+          $('.is-invalid').removeClass('is-invalid')
+          $('.invalid-feedback').remove()
+          
+          setErrorMessages(form, error.responseJSON.errors);
+        } else {
+          showDialog(error.responseJSON)
+        }
+      },
+    })
+  }
 
   function createGudang() {
     let form = $('#crudForm')
@@ -533,7 +565,7 @@
     });
   }
 
-  function cekValidasidelete(Id) {
+  function cekValidasidelete(Id,Aksi) {
     $.ajax({
       url: `{{ config('app.api_url') }}gudang/${Id}/cekValidasi`,
       method: 'POST',
@@ -541,12 +573,28 @@
       beforeSend: request => {
         request.setRequestHeader('Authorization', `Bearer {{ session('access_token') }}`)
       },
+      data:{
+        aksi: Aksi,
+        id: Id
+      },
       success: response => {
         var kondisi = response.kondisi
         if (kondisi == true) {
-          showDialog(response.message['keterangan'])
+          if (!response.editblok) {
+            if (Aksi == 'EDIT') {
+              editGudang(Id)
+            } else {
+              showDialog(response.message['keterangan'])
+            }
+          }else{
+            showDialog(response.message['keterangan'])
+          }
         } else {
-          deleteGudang(Id)
+          if (Aksi=="EDIT") {
+            editGudang(Id)
+          }else if (Aksi=="DELETE"){
+            deleteGudang(Id)
+          }
         }
 
       }
