+++ resources/views/jenisorder/_modal.blade.php
@@ -84,6 +84,7 @@
       let jenisOrderId = form.find('[name=id]').val()
       let action = form.data('action')
       let data = $('#crudForm').serializeArray()
+      var data_id
 
       data.push({
         name: 'sortIndex',
@@ -184,6 +185,8 @@
     setFormBindKeys(form)
 
     activeGrid = null
+    data_id = $('#crudForm').find('[name=id]').val();
+    
 
     form.find('#btnSubmit').prop('disabled', false)
     if (form.data('action') == "view") {
@@ -196,8 +199,39 @@
 
   $('#crudModal').on('hidden.bs.modal', () => {
     activeGrid = '#jqGrid'
+    removeEditingBy(data_id)
+    
     $('#crudModal').find('.modal-body').html(modalBody)
   })
+  function removeEditingBy(id) {
+    $.ajax({
+      url: `{{ config('app.api_url') }}bataledit`,
+      method: 'POST',
+      dataType: 'JSON',
+      headers: {
+        Authorization: `Bearer ${accessToken}`
+      },
+      data: {
+        id: id,
+        aksi: 'BATAL',
+        table: 'jenisorder'
+        
+      },
+      success: response => {
+        $("#crudModal").modal("hide")
+      },
+      error: error => {
+        if (error.status === 422) {
+          $('.is-invalid').removeClass('is-invalid')
+          $('.invalid-feedback').remove()
+          
+          setErrorMessages(form, error.responseJSON.errors);
+        } else {
+          showDialog(error.responseJSON)
+        }
+      },
+    })
+  }
 
   function createJenisOrder() {
     let form = $('#crudForm')
@@ -538,7 +572,7 @@
     })
   }
 
-  function cekValidasidelete(Id) {
+  function cekValidasi(Id,Aksi) {
     $.ajax({
       url: `{{ config('app.api_url') }}jenisorder/${Id}/cekValidasi`,
       method: 'POST',
@@ -546,12 +580,27 @@
       beforeSend: request => {
         request.setRequestHeader('Authorization', `Bearer {{ session('access_token') }}`)
       },
+      data: {
+        aksi: Aksi,
+      },
       success: response => {
         var kondisi = response.kondisi
         if (kondisi == true) {
-          showDialog(response.message['keterangan'])
+          if (!response.editblok) {
+            if (Aksi == 'EDIT') {
+              editJenisOrder(Id)
+            } else {
+              showDialog(response.message)
+            }
+          }else{
+            showDialog(response.message)
+          }
         } else {
-          deleteJenisOrder(Id)
+          if (Aksi=="EDIT") {
+            editJenisOrder(Id)
+          }else if (Aksi=="DELETE"){
+            deleteJenisOrder(Id)
+          }
         }
 
       }
