+++ resources/views/kota/_modal.blade.php
@@ -96,6 +96,7 @@
       let kotaId = form.find('[name=id]').val()
       let action = form.data('action')
       let data = $('#crudForm').serializeArray()
+      var data_id
 
       data.push({
         name: 'sortIndex',
@@ -197,6 +198,7 @@
     setFormBindKeys(form)
 
     activeGrid = null
+    data_id = $('#crudForm').find('[name=id]').val();
 
     form.find('#btnSubmit').prop('disabled', false)
     if (form.data('action') == "view") {
@@ -211,8 +213,39 @@
 
   $('#crudModal').on('hidden.bs.modal', () => {
     activeGrid = '#jqGrid'
+    removeEditingBy(data_id)
     $('#crudModal').find('.modal-body').html(modalBody)
   })
+
+  function removeEditingBy(id) {
+    $.ajax({
+      url: `{{ config('app.api_url') }}bataledit`,
+      method: 'POST',
+      dataType: 'JSON',
+      headers: {
+        Authorization: `Bearer ${accessToken}`
+      },
+      data: {
+        id: id,
+        aksi: 'BATAL',
+        table: 'kota'
+        
+      },
+      success: response => {
+        $("#crudModal").modal("hide")
+      },
+      error: error => {
+        if (error.status === 422) {
+          $('.is-invalid').removeClass('is-invalid')
+          $('.invalid-feedback').remove()
+          
+          setErrorMessages(form, error.responseJSON.errors);
+        } else {
+          showDialog(error.responseJSON)
+        }
+      },
+    })
+  }
 
   function createKota() {
     let form = $('#crudForm')
@@ -548,7 +581,7 @@
     })
   }
 
-  function cekValidasidelete(Id) {
+  function cekValidasidelete(Id,Aksi) {
     $.ajax({
       url: `{{ config('app.api_url') }}kota/${Id}/cekValidasi`,
       method: 'POST',
@@ -556,12 +589,28 @@
       beforeSend: request => {
         request.setRequestHeader('Authorization', `Bearer {{ session('access_token') }}`)
       },
+      data:{
+        aksi: Aksi,
+        id: Id
+      },
       success: response => {
         var kondisi = response.kondisi
         if (kondisi == true) {
-          showDialog(response.message['keterangan'])
+          if (!response.editblok) {
+            if (Aksi == 'EDIT') {
+              editKota(Id)
+            } else {
+              showDialog(response.message['keterangan'])
+            }
+          }else{
+            showDialog(response.message['keterangan'])
+          }
         } else {
-          deleteKota(Id)
+          if (Aksi=="EDIT") {
+            editKota(Id)
+          }else if (Aksi=="DELETE"){
+            deleteKota(Id)
+          }
         }
 
       }
